qgnomeplatform: 0.8.4 -> 0.9.2
hardcode-gsettings patch needed rework due to upstream refactoring; from
the code, it seems that at least part of it may still be necessary,
though I haven’t been able to reproduce the original issue in my own
config. Therefore erring on the side of leaving it in.

qt6-cursor-fix has been merged and appears to be in this release.

Worth noting that qgnomeplatform upstream indicates it is no longer
maintained, and that 0.9.2 is the final release.