bandwhich: add `meta.changelog`
bandwhich has two different changelogs:
1. the in-tree CHANGELOG.md
2. git tags/GitHub Releases

Unfortunately, the CHANGELOG.md contains fewer information from time to
time. One example would be
https://github.com/imsnif/bandwhich/blob/v0.22.0/CHANGELOG.md vs
https://github.com/imsnif/bandwhich/releases/tag/v0.22.0

In this case, the GitHub Release contains a lot more noise due to
various chore(deps) entries.

This is due to
https://github.com/imsnif/bandwhich/blob/b5e48ba9bb466c5e1192755b0235b6b43d1d4966/.github/workflows/require-changelog-for-PRs.yml#L10-L11

I'd argue, the CHANGELOG.md should be preferred here, but in the end
this isn't my package and not my call.