clairvoyant: Add hideUnsupportedVersionBanner option
Clairvoyant v3.1.3 introduced a non-dismissble banner stating that this
version of the app is unsupported, which is shown when the app detects
that it is not running as a Flatpak. This change adds the
"hideUnsupportedVersionBanner" option, disabled by default, which makes
it look like the app is running as a Flatpak.