lxc: fix aarch64 build failure
Starting with commit 1cf2d7357c61b9c9ca5c92cb0e810571edee948a lxc is
failing on aarch64 only. For whatever reason, the version check done
on docbook2man here is no longer registering as >=0.8.8

https://github.com/lxc/lxc/blob/3efa1c3037bff09d12e36727cd6fea2f97fc0a7c/meson.build#L323

This falls through to this line, which changes the behavior of the
configured sgml header.

https://github.com/lxc/lxc/blob/3efa1c3037bff09d12e36727cd6fea2f97fc0a7c/meson.build#L327

Instead of figuring out /why/ this is the case, I'm just going to skip
the check for lxc since we won't ever ship docbook2x < 0.8.