llvmPackages: 16.0.6 -> 17.0.6 on Linux
rustc is already using LLVM 17, so as long as we don't find that
anything widely used part of a NixOS system has a hard requirement of
LLVM 16 over LLVM 17, it makes sense to upgrade the default to
minimise the number of LLVMs in closures.

All my systems build fine with LLVM 17 as the default.