chromium: fix rust toolchain and remove M121 workaround
The rust toolchain is required for chromium since M121.

In the last major bump (M120 -> M121) we had to work around this
requirement because we hadn't had our that part of our toolchain ready.

Until now :)

So this fixes and enables the toolchain for any chromium/electron >= 121
and removes the workaround from the last major bump.