openjdk: 16+36 -> 16.0.2-ga
Build number is set to 1, as this was the last release and thus had no
build number.

Additionally, this update is made despite the fact that the package is
marked as insecure, to pull in the small number of accumulated fixes.

(cherry picked from commit dede7feaa7730c8b43d0f06fd295b865da9a1e6a)