buildMozillaMach: fix vp8/vp9 WebRTC video support
The WebRTC parts of Firefox 122.0 are unconditionally linked against the
vendored libvpx that they ship with This lead to an ABI mismatch¹, when
building with --with-system-libvpx, since our libvpx version differs.

An upstream fix is expected to be shipped in Firefox 122.0.1, which is
due in the next few days.

[1] https://bugzilla.mozilla.org/show_bug.cgi?id=1875201#c11

(cherry picked from commit 8e0ce57c6310792be31f0259963ec75ed57c1ab3)