signald: drop myself (ma27) as maintainer
I only took care of that package because it was essential to operate a
bridge from Matrix to Signal (mautrix-signal). Now that the package uses
libsignal-ffi, this isn't necessary anymore and thus I have no use-case
and no test environment anymore.

I'm still happy to answer questions if possible though, but this doesn't
warrant a maintainer status.