hh-suite: fix build on GCC 13
The build of `hh-suite` stopped working with GCC 13 because
GCC stopped transitively including a couple of headers like
`cstdint` in various scenarios.

There already is an upstream PR proposed that fixes this issue [1]
but hasn't been merged yet. This change pulls in this correction
using `fetchpatch`, fixing the build for now.

[1]: https://github.com/soedinglab/hh-suite/pull/357