buildLinux: structuralize passthru.tests and kernelTests
Present passthru.tests of the result package as an attribute set instead
of a list.

Recommend adding tests by overriding `passthru.tests` with
`overrideAttrs`.

Make the input argument kernelTests accept an attribute set instead of a
list.

Handle the previous override interface of kerneTests as a list with a
warning message and transform the elements into an attribute set in the
form `passthru.tests.extra-kernel-test-<0-based index>`.