nixos/zsh/ohMyZsh: Allow disabling loading oh-my-zsh on a per-user basis
Currently when oh-my-zsh is enabled system-wide, and an individual user also specifies that they want to load another version of oh-my-zsh (via e.g. home-manager), only the system version gets loaded, and not the per-user-overridden version.

This is because when initializing an interactive shell, the files are loaded in the following order:
- `/etc/zshenv` -> we set nothing particular here
- `~/.zshenv` -> here, home-manager sets `$ZSH` if package was overridden by the user
- `/etc/zshrc` -> we set `$ZSH` here, and run `$ZSH/oh-my-zsh.sh`
- `~/.zshrc` -> here, home-manager runs `$ZSH/oh-my-zsh.sh`, sourcing again the system's `$ZSH`, not the user's `$ZSH`, because $ZSH was set by `/etc/zshrc`.

In order to allow individual users to override the oh-my-zsh version that gets loaded (or even disable it entirely and load something different), on a system where it's otherwise enabled globally, and to avoid it being enabled twice if specified in the user's `.zshrc`, we read an additional `$ETC_ZSHRC_DONT_LOAD_OH_MY_ZSH` variable, and if set, we actually don't load `oh-my-zsh`.