python3Packages.gevent: Cherry-pick fixes to events.py for Python < 3.10
In 61dac96af51b77aebb664e18daadd471dbec604f we backported upstream
patch to support Python 3.11, however that patch was not fully tested
and broke compatibility with Python 3.9. In particular, on line 126
of events.py entry_points is not defined.

There were two subsequent commits after that one which fix the issues
in Python < 3.10 branch. Each of them is one line when limited to
events.py, so I backported them too.

This fixes build of python39.pkgs.geventhttpclient in nixos-23.11
branch.