prettier-plugin-sql-cst: init at 0.11.3
Prettier is a well-known code formatter. While it does not have built-in support
for SQL code, there is this plugin SQL-CST (`prettier-plugin-sql-cst`).

Release notes: https://github.com/nene/prettier-plugin-sql-cst/releases

As for the alternatives, there already exists a NixOS package `sql-formatter` as
well as a Node.js package `prettier-plugin-sql`, both of which wrap the same SQL
formatter. However, as the author of this SQL formatter explains, they have
started `prettier-plugin-sql-cst` to improve the original tool (for details, see
https://github.com/sql-formatter-org/sql-formatter?tab=readme-ov-file#the-future).

`pkgs/by-name/po/postlight-parser` has served me as an example for a Yarn
package in Nix.