CODEOWNERS: add raitobezarius on the linux kernel
I'm planning to do some structural changes to the Linux kernel expressions,
for this, I need to be aware of all PRs inflight which will be created.

This is not a permanent code ownership that I'm taking, I don't have the strength
to care seriously about the kernel like others, so I will drop myself at the end of
the work.

Signed-off-by: Raito Bezarius <masterancpp@gmail.com>