luaPackages: Reset upstream source for 'say'
After some time in limbo the LuaRock 'say' is managed by Lunar Modules
and the luarocks.org sources are up to date again. This is important
because it needs to be matched with busted, one of the primary consumers
of this library, which now expects a newer version than is posted on the
Olivine Labs legacy site.