nixos-rebuild: capture the switch output to the journal
Currently nixos-rebuild switch runs switch-to-configuration attached to
the user's terminal, but under systemd-run. This is great and helps take
the switch to completion if the terminal vanishes, but it does not log
the switch output itself to the journal because it is passed straight
through to the terminal. We cannot stop using --pty mode, because some
users do use stdin in their configuration switch scripts.

A while ago I really screwed up my configuration and it killed my entire
desktop session when I switched. There was no log of what it did,
besides guessing at which service changes were its fault while looking
at the journal afterwards.

Fortunately, there is an ancient Unix utility called script(1) (today,
in the util-linux package) which makes digital logs of your
teletypewriter printouts as files, so you can `grep` them without having
to pore over a large stack of folded paper by hand! What a labour
saving!

This is exactly what we want, if the file in question is a pipe to
systemd-cat, and it just takes some clever wrapping with a little bash
script to shell-quote the command losslessly, so we can stick it in
front of the configuration switch.

Arguably logging systemd-run pty mode could be a reasonable feature to
include in systemd, but we already depend on util-linux for
switch-to-configuration and can compose together something that works
today.