xdg-user-dirs: fix darwin builds
A recent update to xdg-utils added xdg-user-dirs as a dep.
Since then we've had a few reports of people running macOS
with xdg-utils in their system for one reason or another
getting a failure on xdg-user-dirs.

I'm not sure it makes high-level sense to use it on macOS,
but it was easy to get it to build...