ollama: 0.1.17 -> 0.1.24
Remove outdated patches and substitution of llama.cpp server.

Patch source build script to remove git.
Ollama's `go generate` script assumes the source is an
initialized git repository, which it is not in nix.
So, remove usage of `git` and replace it with `patch`.
Also replace call to `g++` with `$CXX`.

Run `go generate` to build dependencies using
ollama's custom build script.