python311Packages.pyocr: fix non-Linux builds
Cuneiform doesn't build on non-Linux, causing pyocr to not build on that platform.

The fix was pretty simple; the general idea was to split Cuneiform and Tesseract support, and make Cuneiform support conditional based on the system.

pyocr gracefully degrades when it can't find a tool, so this shouldn't result in adverse behavior.