bat-extras: 2023.06.15 -> 2024.02.12
The most notable change in this version is the fix of `batman` making an
infinite recursion when $BAT_PAGER is set.
See https://github.com/eth-p/bat-extras/releases/tag/v2024.02.12

This commit also changes `sha256` to `hash` and makes the repo name
explicit instead of reusing the package name with `pname`.