picom-next: alias to picom
Since picom v11 there is no more reason to keep a separate picom
derivation that tracks the `next` branch from upstream since the current
release is basically the `next` branch.

Also, picom-next right now is pointing to an old commit, so it is clear
that nobody is maintaining it updated.

If there is a necessity later, we can bring back the picom-next again.