cmake: Fix impurity without sandbox
If a Nix Cmake build occurs without the Nix sandbox on a Debian, Arch,
or Alpine system, CMake may probe /etc, notice the presence of
e.g. /etc/debian_version, and decide that it is running on Ubuntu.

Because it uses the result of this probe to configure the default
value of the CMAKE_INSTALL_LIBDIR variable, this can result in an
impurity where such a build will place built library files in a
different folder than a properly sandboxed / NixOS build.

Fix this by patching out the corresponding logic.