sitespeed-io: unbreak build
This package does not build with nodejs_20 yet.
Yes, this package is heavily outdated (see shortly following commit),
but even with the latest upstream version, it still won't build.

To work around this, we pin the nodejs version down to nodejs_18.

Additionally, makeWrapper isn't needed, as it is already included by
default when using buildNpmPackage.
And python3 isn't needed as nativeBuildInput (but very much needed
further down below in postFixup for wrapping).