sharedown: unbreak build
This is essentially a partial revert of
c271196d42a99c038472994b0fd0bfd9028b4423 which was a successful attempt
at unbreaking this build back in November 2023.

That workaround, however, no longer, well.. works.

Instead I figured out what was wrong with the initial version and fixed
that.
Hopefully this fix lasts longer than the one before.

The important bit that was missing is that `export npm_config_nodedir`
line.

The code in question is copied from
https://nixos.org/manual/nixpkgs/unstable/#javascript-yarn2nix-pitfalls
https://github.com/NixOS/nixpkgs/blob/d176767c02cb2a048e766215078c3d231e666091/doc/languages-frameworks/javascript.section.md#pitfalls-javascript-yarn2nix-pitfalls)

I suggest following the link for even more links as to why some commands
from there are the way they are.