git-branchless: install man pages
This fixes the failure of `git cmd --help` where `cmd` is some
git-branchless alias.

The error is due to a quirk of git: when supplied with the long form
`--help` flag, the git binary eats the flag and instead look for man
page of `cmd`. This errors out unless the man page for `cmd` is properly
installed. In the case of `git-branchless`, this can be achieved with
the command `git-branchless install-man-pages`, as implemented in this
patch.