mqtt-exporter: init at 1.4.1
Two patches (both already merged upstream to master) are needed, one for
providing some basic pyproject properties to allow it to build
successfully.

Second patch addresses the lack of a configuration option to set the
port the metrics exporter should listen on, which is an common option
pre-defined for all exporters.

Signed-off-by: Christoph Heiss <christoph@c8h4.io>