nixos/github-runners: allow removing nix from PATH
In order to not break existing users, we still default to having Nix
in the path, but users can now optionally remove it.

Nix can be considered a huge attack surface, so runners that don't need
it probably shouldn't have it.

In order to realize the added security benefits, the daemon socket is
not accessible when Nix is removed from the path.