ecwolf: add passthru.updateScript
Before this change, nixpkgs-update had to rely on on Repology in order
to determine if there was an ECWolf update. When it comes to ECWolf
updates, Repology is a tertiary source of information. This change
allows nixpkgs-update (and any other tools) to use debian.drdteam.org to
determine if there is an ECWolf update. When it comes to ECWolf updates,
debian.drdteam.org is a primary source of information.