linux: CONFIG_NET_SCH_FQ_CODEL=y
Beginning in 2014, systemd sets the default qdisc to fq_codel,
in order to fight off the general scourge of internet bufferbloat.

https://github.com/systemd/systemd/commit/e6c253e363dee77ef7e5c5f44c4ca55cded3fd47

At the moment this gets built as a module, but since at least 99%
of NixOS users are not going to change their qdisc, let's build
this code into the kernel and save the disk I/O at boot of loading
another kernel module.

Signed-off-by: zzywysm <zzy@zzywysm.com>