Specify guest *system* instead of guest *packages*
… based on feedback from @roberth

This plays nicer with other changes the user might want to make and
also works in conjunction with the `pkgsReadOnly`.

Note that this required deleting the `nixpkgs.pkgs` default because
that's not compatible with setting `nixpkgs.system`.  If `nixpkgs.pkgs`
is defined (even as a default-valued option) then `nixpkgs.system` is
ignored.