nixos/vpp: init
This is a fairly primitive NixOS module, it does not offer RFC42 because
it can be hard to get right on the first try.

Your deployment on your baremetal system may vary and may require personal overrides.
It offers simple defaults for the hugepages setup and kernel modules stuff.

If you want a better NixOS module, please chime in with usecases.

Signed-off-by: Raito Bezarius <masterancpp@gmail.com>