jetbrains-toolbox: fix login flow
This does two things:

Switch to the upstream desktop file. There's no reason to be using our
own, and overriding theirs means we'll miss out on potential changes and
improvements. Their desktop file also registers the application as a
handler for `jetbrains://` URLs, which was previously missing from ours.

Fix the desktop file so that URLs are passed to the application. This
argument was erroneously absent from both their and our `Exec=`, meaning
that even with the scheme handler, it wouldn't actually see any URLs.
`jetbrains://` URLs are used by the application for, among other things,
callbacks during the login flow. Without them, it's not possible to
smoothly log in; instead requiring the user to paste in an auth token.

Change-Id: Ic20469f63c0c3cfbb87f0670afbf5cd68ad7f7c5