atuin: Allow setting database.uri to null
When a password is required to connect to postgres using
services.atuin.database.uri directly would make the password be written
in the nix store, which is suboptimal.
Instead we can have the password in a file accessible only to root by
having systemd read an EnvironmentFile directly, but we must ensure that
this file has priority over the environment set.
Not setting the variable in this case is more straightforward.